Nicer domain search.
If domain doesn't match, then strip off the first component and
search again. Continue till we run out of domain components or
find a match.

For example if foo.bar.baz.box.com is searched for then this leads
to the following searches (in order):
  foo.bar.baz.box.com
  bar.baz.box.com
  baz.box.com
  box.com
The search terminates at the first one which yields a non-empty
result.