Completion for password only input
With the actual behavior if the page only has a password input (password confirmation page...) then it will be filled and overridden by the login afterwards because it's the first of type input. If we change the order than it will correctly work.
I added a comment so that this behavior will not be lost overtime which can easily happened with this kind of details.