POC for a couple of alternative search methods.
Both of these use the same overall pattern:
* Set an environment variable with a program for searching
* Shell out from browserpass to run the search program, passing the
query string as a command arg, and returning results in the same format
as built-in search.

NOTE: the shell command is evaluated like:

`sh -c "$VAR {}"`

Where the {} is replace with the search string provided from the user in
the browser.

The two methods implemented here are:

1. Custom command to find pass entries and then search them
2. Filter results of an empty string query from the built-in search

Details:

*Custom command* - Set the environment variable
BROWSERPASS_SEARCH_COMMAND. This command will be run in a shell, and the
output from stdout will be returned from the DiskStore.Search function -
each line of stdout will be translated to one entry in search results.

An example command:
`BROWSERPASS_SEARCH_COMMAND="find $PASSWORD_STORE_DIR -type f -name *.gpg | sed -e \"s|$PASSWORD_STORE_DIR/\(.*\)\.gpg.*|\1|\" | fzf -f"`

this will walk the password store for files having the `.gpg` extension,
turn them into the name format used by pass/browserpass, and finally
search for the entries using the fzf program.

*Filter* - Set the environment variable BROWSERPASS_FILTER_COMMAND. This
command will be run with the search string as above, however the stdin
will be set to a pipe. The results of running the original search
function with an empty-string query will be sent over the pipe to the
fitler program.

An example command:
`BROWSERPASS_FILTER_COMMAND="fzf -f"`