Add builtin fuzzy support to browserpass #32
This adds builtin support for fuzzy finding in the browserpass search
bar. It also adds options to:

1. Turn on/off the use of fuzzy search from the search bar.
2. Select which fuzzy finding algorithm to use.

The second option allows users to choose between:
* github.com/renstrom/fuzzysearch/fuzzy
* github.com/sahilm/fuzzy

NOTE: the selection of fuzzy algorithm should probably go away before
merging the branch, but it is included right now, since there was a
discussion on #209 about which of those algorithms would be right for
browserpass. This makes switching between the two for evaluation
simpler, until a decision can be made on which to include (or a decision
to keep as-is).